merge symbols and chars
Characters "abc" were handled as []rune similar to Ks character
arrays.
Both types, symbols and character arrays are now merged into one
string type.
The length of a string is always 1. To count the number of runes,
split it first (nyi).
Splitting a string with the empty string, returns a string slice
each holding a single rune converted to string.