cst to numbers or custom types
cst (dyadic $) converts to a custom value of the type of x,
if x has a ConvertTo method.
cst can also convert to basic numbers or slices of basic numeric types
if x is a number of a primitive go type (bool, uint8, ...).
To help, the global variable "int" is predefined as a dict.
@(int@0) / int
@(int@-1) / uint
@(int@2)1 / bool
@(int@8)128 / int8, 8,16,32
@(int@-8)128 / uint8, -8,-16,-32