tail calls
this adds tco for
 g:{$[y;:g[x+y;y-1];x]};g[0;1000000]}
tail calls are detected only with explicit return (:x)
by the parser

at runtime we check if the top is a lambda and if the arguments
match.

lambda function execution is split into 3 parts:
lam0: check for short arguments (project)
lam1: save shadowed args/locals on the stack
lam2: exec and call lam3
lam3: restore shadowed vars

if a tail call matches at runtime lam3 the current lamda,
lam1 the new lambda,
and set the current position/end to the new lambda's code.

currently this is implemented only for multi-argument calls
:f[x;y;z] which use f.(x;y;z) not for :f[x] which is :f@x

this still needs to be done